assets not working in subfolder installs
Im using this extension (and others like it) as a sample on how to do the more complex stuff when coding bolt extensions. Thanks for that!

In my extension, i found out that loading assets with paths as above will not work when bolt is installed in a subfolder. The `$app['resources']->urlPrefix` seems to take care of that. If there is a better way, feel free to point me in the right direction.